+++ src/core/common/instance.hpp
@@ -458,9 +458,14 @@ template <> inline Ip6::Filter &Instance::Get(void)
 }
 
 #if OPENTHREAD_FTD
+template <> inline IndirectSender &Instance::Get(void)
+{
+    return mThreadNetif.mMeshForwarder.mIndirectSender;
+}
+
 template <> inline SourceMatchController &Instance::Get(void)
 {
-    return mThreadNetif.mMeshForwarder.mSourceMatchController;
+    return mThreadNetif.mMeshForwarder.mIndirectSender.mSourceMatchController;
 }
 
 template <> inline AddressResolver &Instance::Get(void)
