+++ src/core/thread/src_match_controller.cpp
@@ -40,6 +40,7 @@
 #include "mac/mac_frame.hpp"
 #include "thread/mesh_forwarder.hpp"
 #include "thread/thread_netif.hpp"
+#include "thread/topology.hpp"
 
 namespace ot {
