+++ src/core/thread/src_match_controller.hpp
@@ -36,10 +36,12 @@
 
 #include "openthread-core-config.h"
 
+#include <openthread/error.h>
 #include "common/locator.hpp"
-#include "thread/topology.hpp"
 
 namespace ot {
+
+class Child;
 
 /**
  * @addtogroup core-source-match-controller
